Lab 3 is done without H

+++ Lab1/g.cpp
@@ -2,10 +2,9 @@
 #include <vector>
 using namespace std;
 
-// Проверка на простоту
 bool isPrime(int n) {
     if (n < 2) return false;
-    for (int i = 2; i * i <= n; i++) { // оптимизация до sqrt(n)
+    for (int i = 2; i * i <= n; i++) {
         if (n % i == 0) return false;
     }
     return true;
@@ -19,7 +18,7 @@ int main() {
     vector<int> superPrimes;     
 
     int number = 2;
-    while ((int)superPrimes.size() < n) {
+    while (superPrimes.size() < n) {
         if (isPrime(number)) {
             primes.push_back(number);
             int index = primes.size();
